+++ lib/src/analyzers/lint_analyzer/reporters/reporters_list/console/lint_console_reporter.dart
@@ -34,7 +34,10 @@ class LintConsoleReporter
     for (final file in records) {
       final lines = [
         ..._reportMetrics('', file.file),
-        ..._reportIssues([...file.issues, ...file.antiPatternCases]),
+        ..._reportIssues(
+          [...file.issues, ...file.antiPatternCases],
+          file.relativePath,
+        ),
         ..._reportEntityMetrics({...file.classes, ...file.functions}),
       ];
 
@@ -46,10 +49,12 @@ class LintConsoleReporter
     }
   }
 
-  Iterable<String> _reportIssues(Iterable<Issue> issues) => (issues.toList()
-        ..sort((a, b) =>
-            a.location.start.offset.compareTo(b.location.start.offset)))
-      .map(_helper.getIssueMessage);
+  Iterable<String> _reportIssues(Iterable<Issue> issues, String relativePath) =>
+      (issues.toList()
+            ..sort((a, b) =>
+                a.location.start.offset.compareTo(b.location.start.offset)))
+          .map((issue) => _helper.getIssueMessage(issue, relativePath))
+          .expand((lines) => lines);
 
   Iterable<String> _reportEntityMetrics(Map<String, Report> reports) =>
       (reports.entries.toList()
@@ -66,9 +71,7 @@ class LintConsoleReporter
             _helper.getMetricReport(metric),
       ];
 
-      return [
-        _helper.getMetricMessage(reportLevel, source, violations),
-      ];
+      return _helper.getMetricMessage(reportLevel, source, violations);
     }
 
     return [];